# TicketManagementSystem/backend/TicketManagementSystem.API/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Management System API
  description: |
    API RESTful para el Sistema de Gestión de Tickets.
    
    ## Autenticación
    Esta API utiliza **JWT Bearer Token** para autenticación. Para acceder a los endpoints protegidos:
    1. Obtener un token mediante `/api/auth/login`
    2. Incluir el token en el header `Authorization: Bearer {token}`
    
    ## Rate Limiting
    - General: 200 requests/minuto
    - Login: 20 requests/minuto
    - Register: 10 requests/minuto
    
    ## Roles
    - **Admin**: Acceso completo al sistema
    - **Agent**: Gestión de tickets asignados
    - **User**: Crear y ver tickets propios
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@ticketmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
  - url: https://localhost:5001/api
    description: Servidor de desarrollo (HTTPS)
  - url: https://api.ticketmanagement.com/api
    description: Servidor de producción
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      description: Autenticar usuario y obtener token JWT
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            example:
              email: "admin@example.com"
              password: "Admin123!"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Demasiadas solicitudes (rate limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/register:
    post:
      tags:
        - Auth
      summary: Registrar nuevo usuario
      description: Crear una nueva cuenta de usuario
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
            example:
              email: "newuser@example.com"
              password: "SecurePass123!"
              fullName: "Juan Pérez"
              role: "User"
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseDto'
        '400':
          description: Datos de entrada inválidos o email ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Demasiadas solicitudes (rate limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refrescar token JWT
      description: Obtener un nuevo token usando el refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequestDto'
      responses:
        '200':
          description: Token refrescado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponseDto'
        '401':
          description: Refresh token inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Cerrar sesión
      description: Invalidar el token actual
      operationId: logout
      responses:
        '200':
          description: Sesión cerrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponseDto'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ==================== TICKETS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido mediante el endpoint `/api/auth/login`.
        
        Formato: `Bearer {token}`
        
        Ejemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email del usuario
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 6
          description: Contraseña del usuario
          example: "SecurePass123!"
    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInJvbGUiOiJBZG1pbiIsImV4cCI6MTczMjYyMDAwMH0.signature"
        refreshToken:
          type: string
          description: Token para refrescar el access token
          example: "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
        expiresAt:
          type: string
          format: date-time
          description: Fecha de expiración del token
          example: "2025-11-25T16:00:00Z"
        user:
          $ref: '#/components/schemas/UserBasicDto'
    RegisterDto:
      type: object
      required:
        - email
        - password
        - fullName
        - role
      properties:
        email:
          type: string
          format: email
          description: Email único del usuario
          example: "newuser@example.com"
        password:
          type: string
          format: password
          minLength: 6
          description: Contraseña (mínimo 6 caracteres)
          example: "SecurePass123!"
        fullName:
          type: string
          maxLength: 100
          description: Nombre completo del usuario
          example: "Juan Pérez"
        role:
          type: string
          enum: [Admin, Agent, User]
          default: User
          description: Rol del usuario
          example: "User"
    RegisterResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de confirmación
          example: "Usuario registrado exitosamente"
        user:
          $ref: '#/components/schemas/UserBasicDto'
    RefreshTokenRequestDto:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token válido
          example: "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
    RefreshTokenResponseDto:
      type: object
      properties:
        token:
          type: string
          description: Nuevo JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Nuevo refresh token
          example: "bmV3IHJlZnJlc2ggdG9rZW4..."
        expiresAt:
          type: string
          format: date-time
          description: Nueva fecha de expiración
          example: "2025-11-25T17:00:00Z"
    LogoutResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de confirmación
          example: "Sesión cerrada exitosamente"

    # ==================== USER SCHEMAS ====================
    UserBasicDto:
      type: object
      properties:
        id:
          type: integer
          description: ID único del usuario
          example: 1
        email:
          type: string
          format: email
          description: Email del usuario
          example: "user@example.com"
        fullName:
          type: string
          description: Nombre completo
          example: "Juan Pérez"
        role:
          type: string
          enum: [Admin, Agent, User]
          description: Rol del usuario
          example: "User"
...
    ProblemDetails:
      type: object
      description: Formato de error según RFC 7807
      properties:
        type:
          type: string
          description: URI que identifica el tipo de error
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          description: Título del error
          example: "Bad Request"
        status:
          type: integer
          description: Código HTTP
          example: 400
        detail:
          type: string
          description: Descripción detallada del error
          example: "El campo 'email' es requerido"
        instance:
          type: string
          description: URI de la instancia del error
          example: "/api/auth/login"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Errores de validación por campo
          example:
            email:
              - "El campo 'email' es requerido"
              - "El formato del email es inválido"
            password:
              - "La contraseña debe tener al menos 6 caracteres"
        traceId:
          type: string
          description: ID de traza para debugging
          example: "00-abc123def456-789xyz-00"